+++ pkgs/watcher/test/directory_watcher/link_tests.dart
@@ -32,6 +32,7 @@ void _linkTests({required bool isNative}) {
     createDir('targets');
     createDir('links');
     writeFile('targets/a.target');
+    sleepUntilNewModificationTime();
     writeFile('targets/b.target');
     writeLink(link: 'links/a.link', target: 'targets/a.target');
     await startWatcher(path: 'links');
@@ -46,7 +47,7 @@ void _linkTests({required bool isNative}) {
       'notifies when a link is replaced with a link to a different target '
       'with different contents', () async {
     writeFile('targets/a.target', contents: 'a');
-    writeFile('targets/b.target', contents: 'b');
+    writeFile('targets/b.target', contents: 'ab');
     writeLink(link: 'links/a.link', target: 'targets/a.target');
     await startWatcher(path: 'links');