+++ pkgs/watcher/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 1.1.5-wip
 
+- Polling watchers now check file sizes as well as "last modified" times, so
+  they are less likely to miss changes on platforms with low resolution
+  timestamps.
 - Bug fix: with `FileWatcher` on MacOS, a modify event was sometimes reported if
   the file was created immediately before the watcher was created. Now, if the
   file exists when the watcher is created then this modify event is not sent.